Update 0.20 Gráfica implementada ( Apenas JOptionPane ) Necessario corrigir op 1

+++ src/main/NumericAlgorithm.kt
@@ -2,44 +2,41 @@ package main
 
 import java.lang.Math.pow
 
-class NumericAlgorithm (description: String, title: String, author: String, version : Int): Algorithm(description,title,author,version)
-{
+class NumericAlgorithm(description: String, title: String, author: String, version: Int) : Algorithm(description, title, author, version) {
 
     private val map: String = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
 
+
     fun shortenerEncode(inputNumber: Int): String {
 
-        var inputNumber: Int = inputNumber
-        var newBase =  StringBuilder()
+        var inputNumber: Int = inputNumber.toInt()
+        var newBase = StringBuilder()
 
 
-        while(inputNumber > 0)
-        {
-            newBase.append(map[inputNumber%62])
-            inputNumber/=62
+        while (inputNumber > 0) {
+            newBase.append(map[inputNumber % 62])
+            inputNumber /= 62
         }
 
         return newBase.reverse().toString()
     }
 
 
-    fun shortenerDecode (inputNumber: String): Int
-    {
+    fun shortenerDecode(inputNumber: String): Int {
         var inputNumber: String = inputNumber.reversed()
         var tam = 0
         var newNum: Int = 0
         var aux = 0
 
-        while (tam < inputNumber.length)
-        {
+        while (tam < inputNumber.length) {
 
-            for(indice in map.indices)
+            for (indice in map.indices)
                 if (inputNumber[tam] == map[indice])
-                     aux = indice
+                    aux = indice
 
 
-             newNum += pow(62.0, tam.toDouble()).toInt() * aux
-             tam++
+            newNum += pow(62.0, tam.toDouble()).toInt() * aux
+            tam++
 
         }
 
@@ -47,19 +44,14 @@ class NumericAlgorithm (description: String, title: String, author: String, vers
     }
 
 
-
-
     fun generateNumberSequence(inputText: String): Int {
 
         var inputText: String = inputText
         var h: Int = inputText.hashCode()
-        val M = pow(2.0, 32.0)- 1
+        val M = pow(2.0, 32.0) - 1
 
         return ((h and 0x7fffffff) % M).toInt()
     }
-   
-
-
 
 
 }
